Draft improvements.
