+++ .gitignore
@@ -14,7 +14,6 @@ Thumbs.db
 
 # composer vendor dir
 /vendor
-
 /composer.lock
 
 # composer itself is not needed
@@ -33,3 +32,5 @@ phpunit.phar
 # Phan
 analysis.txt
 
+#node_modules
+/node_modules
