+++ src/AssetManager.php
@@ -4,41 +4,29 @@ declare(strict_types=1);
 namespace Yiisoft\Assets;
 
 use Psr\Log\LoggerInterface;
-use Yiisoft\Aliases\Aliases;
 use Yiisoft\Assets\Exception\InvalidConfigException;
 
 /**
  * AssetManager manages asset bundle configuration and loading.
  *
- * AssetManager is configured in config/web.php. You can access that instance via $container->get(AssetManager::class).
+ * Configured in config/common.php. You can access that instance via $container->get(AssetManager::class).
  *
  * You can modify its configuration by adding an array to your application config under `components` as shown in the
  * following example:
  *
  * ```php
  * AssetManager::class => function (ContainerInterface $container) {
- *     $aliases = $container->get(Aliases::class);
- *     $assetConverterInterface = $container->get(AssetConverterInterface::class);
- *     $fileSystem = $container->get(Filesystem::class);
- *     $logger = $container->get(LoggerInterface::class);
+ *     $assetManager = new AssetManager($container->get(LoggerInterface::class));
  *
- *     $assetManager = new AssetManager($fileSystem, $logger);
- *
- *     $assetManager->setBasePath($aliases->get('@basePath'));
- *     $assetManager->setBaseUrl($aliases->get('@baseUrl'));
- *     $assetManager->setConverter($assetConverterInterface);
- *
+ *     $assetManager->setConverter($container->get(AssetConverterInterface::class));
+ *     $assetManager->setPublisher($container->get(AssetPublisherInterface::class));
+
  *     return $assetManager;
  * },
  * ```
  */
 final class AssetManager
 {
-    /**
-     * @var Aliases
-     */
-    private Aliases $aliases;
-
     /**
      * @var array AssetBundle[] list of the registered asset bundles. The keys are the bundle names, and the values
      * are the registered {@see AssetBundle} objects.
@@ -48,51 +36,16 @@ final class AssetManager
     private array $assetBundles = [];
 
     /**
-     * @var bool whether to append a timestamp to the URL of every published asset. When this is true, the URL of a
-     * published asset may look like `/path/to/asset?v=timestamp`, where `timestamp` is the last modification time of
-     * the published asset file. You normally would want to set this property to true when you have enabled HTTP caching
-     * for assets, because it allows you to bust caching when the assets are updated.
+     * AssetConverter component.
+     *
+     * @var AssetConverterInterface $converter
      */
-    private bool $appendTimestamp = false;
-
-    /**
-     * @var array mapping from source asset files (keys) to target asset files (values).
-     *
-     * This property is provided to support fixing incorrect asset file paths in some asset bundles. When an asset
-     * bundle is registered with a view, each relative asset file in its {@see AssetBundle::css|css} and
-     * {@see AssetBundle::js|js} arrays will be examined against this map. If any of the keys is found to be the last
-     * part of an asset file (which is prefixed with {@see AssetBundle::sourcePath} if available), the corresponding
-     * value will replace the asset and be registered with the view. For example, an asset file `my/path/to/jquery.js`
-     * matches a key `jquery.js`.
-     *
-     * Note that the target asset files should be absolute URLs, domain relative URLs (starting from '/') or paths
-     * relative to {@see baseUrl} and {@see basePath}.
-     *
-     * In the following example, any assets ending with `jquery.min.js` will be replaced with `jquery/dist/jquery.js`
-     * which is relative to {@see baseUrl} and {@see basePath}.
-     *
-     * ```php
-     * [
-     *     'jquery.min.js' => 'jquery/dist/jquery.js',
-     * ]
-     * ```
-     */
-    private array $assetMap = [];
+    private AssetConverterInterface $converter;
 
     /**
      * @var AssetPublisher published assets
      */
     private AssetPublisher $publisher;
-
-    /**
-     * @var string|null the root directory storing the published asset files.
-     */
-    private ?string $basePath = null;
-
-    /**
-     * @var string|null the base URL through which the published asset files can be accessed.
-     */
-    private ?string $baseUrl = null;
 
     /**
      * @var array list of asset bundle configurations. This property is provided to customize asset bundles.
@@ -109,13 +62,6 @@ final class AssetManager
     private array $bundles = [];
 
     /**
-     * AssetConverter component.
-     *
-     * @var AssetConverterInterface $converter
-     */
-    private AssetConverterInterface $converter;
-
-    /**
      * @var array the registered CSS files.
      *
      * {@see registerCssFile()}
@@ -123,73 +69,9 @@ final class AssetManager
     private array $cssFiles = [];
 
     /**
-     * @var int the permission to be set for newly generated asset directories. This value will be used by PHP chmod()
-     * function. No umask will be applied. Defaults to 0775, meaning the directory is read-writable by owner
-     * and group, but read-only for other users.
-     */
-    private int $dirMode = 0775;
-
-    /**
      * @var array $dummyBundles
      */
     private array $dummyBundles;
-
-    /**
-     * @var int the permission to be set for newly published asset files. This value will be used by PHP chmod()
-     * function. No umask will be applied. If not set, the permission will be determined by the current
-     * environment.
-     */
-    private int $fileMode = 0755;
-
-    /**
-     * @var bool whether the directory being published should be copied even if it is found in the target directory.
-     * This option is used only when publishing a directory. You may want to set this to be `true` during the
-     * development stage to make sure the published directory is always up-to-date. Do not set this to true
-     * on production servers as it will significantly degrade the performance.
-     */
-    private bool $forceCopy = false;
-
-    /**
-     * @var callable a callback that will be called to produce hash for asset directory generation. The signature of the
-     * callback should be as follows:
-     *
-     * ```
-     * function ($path)
-     * ```
-     *
-     * where `$path` is the asset path. Note that the `$path` can be either directory where the asset files reside or a
-     * single file. For a CSS file that uses relative path in `url()`, the hash implementation should use the directory
-     * path of the file instead of the file path to include the relative asset files in the copying.
-     *
-     * If this is not set, the asset manager will use the default CRC32 and filemtime in the `hash` method.
-     *
-     * Example of an implementation using MD4 hash:
-     *
-     * ```php
-     * function ($path) {
-     *     return hash('md4', $path);
-     * }
-     * ```
-     */
-    private $hashCallback;
-
-    /**
-     * @var bool whether to use symbolic link to publish asset files. Defaults to false, meaning asset files are copied
-     * to {@see basePath}. Using symbolic links has the benefit that the published assets will always be
-     * consistent with the source assets and there is no copy operation required. This is especially useful
-     * during development.
-     *
-     * However, there are special requirements for hosting environments in order to use symbolic links. In particular,
-     * symbolic links are supported only on Linux/Unix, and Windows Vista/2008 or greater.
-     *
-     * Moreover, some Web servers need to be properly configured so that the linked assets are accessible to Web users.
-     * For example, for Apache Web server, the following configuration directive should be added for the Web folder:
-     *
-     * ```apache
-     * Options FollowSymLinks
-     * ```
-     */
-    private bool $linkAssets = false;
 
     /**
      * @var array the registered JS files.
@@ -198,27 +80,14 @@ final class AssetManager
      */
     private array $jsFiles = [];
 
+    /**
+     * @var LoggerInterface $logger
+     */
     private LoggerInterface $logger;
 
-    public function __construct(Aliases $aliases, LoggerInterface $logger)
+    public function __construct(LoggerInterface $logger)
     {
-        $this->aliases = $aliases;
         $this->logger = $logger;
-    }
-
-    public function getAliases(): Aliases
-    {
-        return $this->aliases;
-    }
-
-    public function getAssetMap(): array
-    {
-        return $this->assetMap;
-    }
-
-    public function getAppendTimestamp(): bool
-    {
-        return $this->appendTimestamp;
     }
 
     /**
@@ -231,24 +100,6 @@ final class AssetManager
         return $this->assetBundles;
     }
 
-    public function getBasePath(): ?string
-    {
-        if (!empty($this->basePath)) {
-            $this->basePath = $this->aliases->get($this->basePath);
-        }
-
-        return $this->basePath;
-    }
-
-    public function getBaseUrl(): ?string
-    {
-        if (!empty($this->baseUrl)) {
-            $this->baseUrl = $this->aliases->get($this->baseUrl);
-        }
-
-        return $this->baseUrl;
-    }
-
     /**
      * Returns the named asset bundle.
      *
@@ -259,13 +110,12 @@ final class AssetManager
      *
      * @return AssetBundle the asset bundle instance
      *
-     * @throws \InvalidArgumentException
      * @throws InvalidConfigException
      */
     public function getBundle(string $name): AssetBundle
     {
         if (!isset($this->bundles[$name])) {
-            return $this->bundles[$name] = $this->publisher->loadBundle($this, $name, []);
+            return $this->bundles[$name] = $this->publisher->loadBundle($name, []);
         }
 
         if ($this->bundles[$name] instanceof AssetBundle) {
@@ -273,18 +123,18 @@ final class AssetManager
         }
 
         if (\is_array($this->bundles[$name])) {
-            return $this->bundles[$name] = $this->publisher->loadBundle($this, $name, $this->bundles[$name]);
+            return $this->bundles[$name] = $this->publisher->loadBundle($name, $this->bundles[$name]);
         }
 
         if ($this->bundles[$name] === false) {
             return $this->loadDummyBundle($name);
         }
 
-        throw new \InvalidArgumentException("Invalid asset bundle configuration: $name");
+        throw new InvalidConfigException("Invalid asset bundle configuration: $name");
     }
 
     /**
-     * Returns the asset converter.
+     * Returns the AssetConverterInterface
      *
      * @return AssetConverterInterface the asset converter.
      */
@@ -293,115 +143,38 @@ final class AssetManager
         return $this->converter;
     }
 
+    /**
+     * Return config array CSS AssetBundle.
+     *
+     * @return array
+     */
     public function getCssFiles(): array
     {
         return $this->cssFiles;
     }
 
-    public function getDirMode(): int
-    {
-        return $this->dirMode;
-    }
-
-    public function getFileMode(): int
-    {
-        return $this->fileMode;
-    }
-
-    public function getForceCopy(): bool
-    {
-        return $this->forceCopy;
-    }
-
+    /**
+     * Return config array JS AssetBundle.
+     *
+     * @return array
+     */
     public function getJsFiles(): array
     {
         return $this->jsFiles;
     }
 
-    public function getLinkAssets(): bool
-    {
-        return $this->linkAssets;
-    }
-
-    public function getHashCallback(): ?callable
-    {
-        return $this->hashCallback;
-    }
-
-    public function getPublish(): AssetPublisher
+    /**
+     * Return the AssetPublisherInterface
+     *
+     * @return AssetPublisherInterface
+     */
+    public function getPublisher(): AssetPublisherInterface
     {
         return $this->publisher;
     }
 
-    public function getPublishedPath(?string $sourcePath): ?string
-    {
-        return $this->publisher->getPublishedPath($sourcePath);
-    }
-
-    public function getPublishedUrl(?string $sourcePath): ?string
-    {
-        return $this->publisher->getPublishedUrl($sourcePath);
-    }
-
     /**
-     * Set appendTimestamp.
-     *
-     * @param bool $value
-     *
-     * @return void
-     *
-     * {@see appendTimestamp}
-     */
-    public function setAppendTimestamp(bool $value): void
-    {
-        $this->appendTimestamp = $value;
-    }
-
-    /**
-     * Set assetMap.
-     *
-     * @param array $value
-     *
-     * @return void
-     *
-     * {@see assetMap}
-     */
-    public function setAssetMap(array $value): void
-    {
-        $this->assetMap = $value;
-    }
-
-    /**
-     * Set basePath.
-     *
-     * @param string|null $value
-     *
-     * @return void
-     *
-     * {@see basePath}
-     */
-    public function setBasePath(?string $value): void
-    {
-        $this->basePath = $value;
-    }
-
-    /**
-     * Set baseUrl.
-     *
-     * @param string|null $value
-     *
-     * @return void
-     *
-     * {@see baseUrl}
-     */
-    public function setBaseUrl(?string $value): void
-    {
-        $this->baseUrl = $value;
-    }
-
-
-    /**
-     * Set bundles.
+     * This property is provided to customize asset bundles.
      *
      * @param array $value
      *
@@ -415,11 +188,10 @@ final class AssetManager
     }
 
     /**
-     * Sets the asset converter.
+     * AssetConverter component.
      *
      * @param AssetConverterInterface $value the asset converter. This can be eitheran object implementing the
-     * {@see AssetConverterInterface}, or a configuration array that can be used
-     * to create the asset converter object.
+     * {@see AssetConverterInterface}, or a configuration array that can be used to create the asset converter object.
      */
     public function setConverter(AssetConverterInterface $value): void
     {
@@ -427,21 +199,7 @@ final class AssetManager
     }
 
     /**
-     * Set hashCallback.
-     *
-     * @param callable $value
-     *
-     * @return void
-     *
-     * {@see hashCallback}
-     */
-    public function setHashCallback(callable $value): void
-    {
-        $this->hashCallback = $value;
-    }
-
-    /**
-     * Set publisher.
+     * AssetPublisher component.
      *
      * @param AssetPublisher $value
      *
@@ -449,11 +207,19 @@ final class AssetManager
      *
      * {@see publisher}
      */
-    public function setPublisher(AssetPublisher $value): void
+    public function setPublisher(AssetPublisherInterface $value): void
     {
         $this->publisher = $value;
     }
 
+    /**
+     * Generate the array configuration of the AssetBundles
+     *
+     * @param array $names
+     * @param integer|null $position
+     *
+     * @return void
+     */
     public function register(array $names, ?int $position = null): void
     {
         foreach ($names as $name) {
@@ -470,14 +236,7 @@ final class AssetManager
      * and {@see registerAssetBundle()} instead.
      *
      * @param string $url the CSS file to be registered.
-     * @param array $options the HTML attributes for the link tag. Please refer to {@see \Yiisoft\Html\Html::cssFile()}
-     * for the supported options. The following options are specially handled and are not treated as HTML
-     * attributes:
-     *
-     *   - `depends`: array, specifies the names of the asset bundles that this CSS file depends on.
-     *
-     * @param string $key the key that identifies the CSS script file. If null, it will use $url as the key. If two CSS
-     * files are registered with the same key, the latter will overwrite the former.
+     * @param array $options the HTML attributes for the link tag.
      *
      * @return void
      */
@@ -500,18 +259,10 @@ final class AssetManager
      * @param array $options the HTML attributes for the script tag. The following options are specially handled and
      * are not treated as HTML attributes:
      *
-     * - `depends`: array, specifies the names of the asset bundles that this JS file depends on.
      * - `position`: specifies where the JS script tag should be inserted in a page. The possible values are:
-     *     * [[POS_HEAD]]: in the head section
-     *     * [[POS_BEGIN]]: at the beginning of the body section
-     *     * [[POS_END]]: at the end of the body section. This is the default value.
-     *
-     * Please refer to {@see \Yiisoft\Html\Html::jsFile()} for other supported options.
-     *
-     * @param string $key the key that identifies the JS script file. If null, it will use $url as the key. If two JS
-     * files are registered with the same key at the same position, the latter will overwrite the former.
-     * Note that position option takes precedence, thus files registered with the same key, but different
-     * position option will not override each other.
+     *     * {@see \Yiisoft\View\WebView::POSITION_HEAD} in the head section
+     *     * {@see \Yiisoft\View\WebView::POSITION_BEGIN} at the beginning of the body section
+     *     * {@see \Yiisoft\View\WebView::POSITION_END} at the end of the body section. This is the default value.
      *
      * @return void
      */
@@ -525,6 +276,76 @@ final class AssetManager
 
         $this->jsFiles[$key]['url'] = $url;
         $this->jsFiles[$key]['attributes'] = $options;
+    }
+
+    /**
+     * Converter files SASS, SCSS, Stylus to CSS.
+     *
+     * @param AssetBundle $bundle
+     *
+     * @return AssetBundle
+     */
+    private function converterCss(AssetBundle $bundle): AssetBundle
+    {
+        foreach ($bundle->css as $i => $css) {
+            if (\is_array($css)) {
+                $file = \array_shift($css);
+                if (AssetUtil::isRelative($file)) {
+                    $css = \array_merge($bundle->cssOptions, $css);
+
+                    if (is_file("$bundle->basePath/$file")) {
+                        \array_unshift($css, $this->converter->convert(
+                            $file,
+                            $bundle->basePath,
+                            $bundle->converterOptions
+                        ));
+
+                        $bundle->css[$i] = $css;
+                    }
+                }
+            } elseif (AssetUtil::isRelative($css)) {
+                if (is_file("$bundle->basePath/$css")) {
+                    $bundle->css[$i] = $this->converter->convert($css, $bundle->basePath, $bundle->converterOptions);
+                }
+            }
+        }
+
+        return $bundle;
+    }
+
+    /**
+     * Converter files CoffeScript, TypeScript to JavaScript.
+     *
+     * @param AssetBundle $bundle
+     *
+     * @return AssetBundle
+     */
+    private function converterJs(AssetBundle $bundle): AssetBundle
+    {
+        foreach ($bundle->js as $i => $js) {
+            if (\is_array($js)) {
+                $file = \array_shift($js);
+                if (AssetUtil::isRelative($file)) {
+                    $js = \array_merge($bundle->jsOptions, $js);
+
+                    if (is_file("$bundle->basePath/$file")) {
+                        \array_unshift($js, $this->converter->convert(
+                            $file,
+                            $bundle->basePath,
+                            $bundle->converterOptions
+                        ));
+
+                        $bundle->js[$i] = $js;
+                    }
+                }
+            } elseif (AssetUtil::isRelative($js)) {
+                if (is_file("$bundle->basePath/$js")) {
+                    $bundle->js[$i] = $this->converter->convert($js, $bundle->basePath);
+                }
+            }
+        }
+
+        return $bundle;
     }
 
     /**
@@ -591,12 +412,11 @@ final class AssetManager
      * @param string $name AssetBunle name class.
      *
      * @return AssetBundle
-     * @throws InvalidConfigException
      */
     private function loadDummyBundle(string $name): AssetBundle
     {
         if (!isset($this->dummyBundles[$name])) {
-            $this->dummyBundles[$name] = $this->publisher->loadBundle($this, $name, [
+            $this->dummyBundles[$name] = $this->publisher->loadBundle($name, [
                 'sourcePath' => null,
                 'js' => [],
                 'css' => [],
@@ -607,6 +427,13 @@ final class AssetManager
         return $this->dummyBundles[$name];
     }
 
+    /**
+     * Process CSS, JS files.
+     *
+     * @param string $name
+     *
+     * @return void
+     */
     private function registerFiles(string $name): void
     {
         if (!isset($this->assetBundles[$name])) {
@@ -619,6 +446,42 @@ final class AssetManager
             $this->registerFiles($dep);
         }
 
-        $this->publisher->registerAssetFiles($this, $bundle);
+        $this->registerAssetFiles($bundle);
+    }
+
+    /**
+     * Registers the CSS and JS files with the given view.
+     *
+     * @param AssetBundle $bundle the asset files are to be registered in the view.
+     *
+     * @return void
+     */
+    private function registerAssetFiles(AssetBundle $bundle): void
+    {
+        if (isset($bundle->basePath, $bundle->baseUrl) && !empty($this->converter)) {
+            $this->converterCss($bundle);
+            $this->converterJs($bundle);
+        }
+
+        foreach ($bundle->js as $js) {
+            if (\is_array($js)) {
+                $file = array_shift($js);
+                $ext = pathinfo($file);
+                $options = array_merge($bundle->jsOptions, $js);
+                $this->registerJsFile($this->publisher->getAssetUrl($bundle, $file), $options);
+            } elseif ($js !== null) {
+                $this->registerJsFile($this->publisher->getAssetUrl($bundle, $js), $bundle->jsOptions);
+            }
+        }
+
+        foreach ($bundle->css as $css) {
+            if (\is_array($css)) {
+                $file = array_shift($css);
+                $options = array_merge($bundle->cssOptions, $css);
+                $this->registerCssFile($this->publisher->getAssetUrl($bundle, $file), $options);
+            } elseif ($css !== null) {
+                $this->registerCssFile($this->publisher->getAssetUrl($bundle, $css), $bundle->cssOptions);
+            }
+        }
     }
 }
