+++ tests/AssetManagerTest.php
@@ -3,8 +3,11 @@ declare(strict_types=1);
 
 namespace Yiisoft\Assets\Tests;
 
+use Yiisoft\Assets\AssetBundle;
 use Yiisoft\Assets\AssetConverterInterface;
 use Yiisoft\Assets\Tests\stubs\JqueryAsset;
+use Yiisoft\Assets\Tests\stubs\Level3Asset;
+use Yiisoft\Assets\Tests\stubs\PositionAsset;
 use Yiisoft\Assets\Tests\stubs\SourceAsset;
 
 final class AssetManagerTest extends TestCase
@@ -17,20 +20,20 @@ final class AssetManagerTest extends TestCase
         );
     }
 
-    public function testGetPublishedPath(): void
+    public function testGetPublishedPathLinkAssetsFalse(): void
     {
         $bundle = new SourceAsset();
 
         $sourcePath = $this->aliases->get($bundle->sourcePath);
 
         $path = $sourcePath . @filemtime($sourcePath);
-        $path = sprintf('%x', crc32($path . '|' . $this->assetManager->getLinkAssets()));
+        $path = sprintf('%x', crc32($path . '|' . $this->assetManager->getPublisher()->getLinkAssets()));
 
         $this->assertEmpty($this->assetManager->getAssetBundles());
         $this->assetManager->register([SourceAsset::class]);
 
         $this->assertEquals(
-            $this->assetManager->getPublishedPath($bundle->sourcePath),
+            $this->assetManager->getPublisher()->getPublishedPath($bundle->sourcePath),
             $this->aliases->get("@public/assets/$path")
         );
     }
@@ -41,7 +44,7 @@ final class AssetManagerTest extends TestCase
 
         $this->assetManager->register([SourceAsset::class]);
 
-        $this->assertNull($this->assetManager->getPublishedPath('/wrong'));
+        $this->assertNull($this->assetManager->getPublisher()->getPublishedPath('/wrong'));
     }
 
     public function testGetPublishedUrl(): void
@@ -51,14 +54,14 @@ final class AssetManagerTest extends TestCase
         $sourcePath = $this->aliases->get($bundle->sourcePath);
 
         $path = $sourcePath . @filemtime($sourcePath);
-        $path = sprintf('%x', crc32($path . '|' . $this->assetManager->getLinkAssets()));
+        $path = sprintf('%x', crc32($path . '|' . $this->assetManager->getPublisher()->getLinkAssets()));
 
         $this->assertEmpty($this->assetManager->getAssetBundles());
 
         $this->assetManager->register([SourceAsset::class]);
 
         $this->assertEquals(
-            $this->assetManager->getPublishedUrl($bundle->sourcePath),
+            $this->assetManager->getPublisher()->getPublishedUrl($bundle->sourcePath),
             "/baseUrl/$path"
         );
     }
@@ -69,7 +72,7 @@ final class AssetManagerTest extends TestCase
 
         $this->assetManager->register([SourceAsset::class]);
 
-        $this->assertNull($this->assetManager->getPublishedUrl('/wrong'));
+        $this->assertNull($this->assetManager->getPublisher()->getPublishedUrl('/wrong'));
     }
 
     public function testAssetManagerSetBundles(): void
@@ -109,29 +112,156 @@ final class AssetManagerTest extends TestCase
         );
     }
 
-    public function testAssetManagerSetAssetMap(): void
+    /**
+     * @return array
+     */
+    public function positionProvider(): array
     {
-        $urlJs = '//testme.css';
+        return [
+            [1, true],
+            [1, false],
+            [2, true],
+            [2, false],
+            [3, true],
+            [3, false],
+        ];
+    }
 
-        $this->assetManager->setAssetMap(
-            [
-                'jquery.js' => $urlJs,
+    /**
+     * @dataProvider positionProvider
+     *
+     * @param int $pos
+     * @param bool $jqAlreadyRegistered
+     */
+    public function testPositionDependency(int $pos, bool $jqAlreadyRegistered): void
+    {
+        $this->assetManager->setBundles([
+            PositionAsset::class => [
+                'jsOptions' =>  [
+                    'position' => $pos,
+                ],
             ]
-        );
+        ]);
 
         $this->assertEmpty($this->assetManager->getAssetBundles());
 
-        $this->assetManager->register([JqueryAsset::class]);
+        if ($jqAlreadyRegistered) {
+            $this->assetManager->register([JqueryAsset::class, PositionAsset::class]);
+        } else {
+            $this->assetManager->register([PositionAsset::class]);
+        }
 
-        $this->assertStringContainsString(
-            $urlJs,
-            $this->assetManager->getJsFiles()[$urlJs]['url']
+        $this->assertCount(3, $this->assetManager->getAssetBundles());
+        $this->assertArrayHasKey(PositionAsset::class, $this->assetManager->getAssetBundles());
+        $this->assertArrayHasKey(JqueryAsset::class, $this->assetManager->getAssetBundles());
+        $this->assertArrayHasKey(Level3Asset::class, $this->assetManager->getAssetBundles());
+
+        $this->assertInstanceOf(
+            AssetBundle::class,
+            $this->assetManager->getAssetBundles()[PositionAsset::class]
+        );
+        $this->assertInstanceOf(
+            AssetBundle::class,
+            $this->assetManager->getAssetBundles()[JqueryAsset::class]
+        );
+        $this->assertInstanceOf(
+            AssetBundle::class,
+            $this->assetManager->getAssetBundles()[Level3Asset::class]
+        );
+
+        $this->assertArrayHasKey(
+            'position',
+            $this->assetManager->getAssetBundles()[PositionAsset::class]->jsOptions
+        );
+        $this->assertEquals(
+            $pos,
+            $this->assetManager->getAssetBundles()[PositionAsset::class]->jsOptions['position']
+        );
+        $this->assertArrayHasKey(
+            'position',
+            $this->assetManager->getAssetBundles()[JqueryAsset::class]->jsOptions
+        );
+
+        $this->assertEquals(
+            $pos,
+            $this->assetManager->getAssetBundles()[JqueryAsset::class]->jsOptions['position']
+        );
+        $this->assertArrayHasKey(
+            'position',
+            $this->assetManager->getAssetBundles()[Level3Asset::class]->jsOptions
+        );
+        $this->assertEquals(
+            $pos,
+            $this->assetManager->getAssetBundles()[Level3Asset::class]->jsOptions['position']
+        );
+
+        $this->assertEquals(
+            [
+                'position' => $pos
+            ],
+            $this->assetManager->getJsFiles()['/js/jquery.js']['attributes']
         );
         $this->assertEquals(
             [
-                'position' => 3
+                'position' => $pos
             ],
-            $this->assetManager->getJsFiles()[$urlJs]['attributes']
+            $this->assetManager->getJsFiles()['/files/jsFile.js']['attributes']
         );
+
+    }
+
+    /**
+     * @return array
+     */
+    public function positionProvider2(): array
+    {
+        return [
+            [1, true],
+            [1, false],
+            [2, true],
+            [2, false],
+        ];
+    }
+
+    /**
+     * @dataProvider positionProvider2
+     *
+     * @param int $pos
+     * @param bool $jqAlreadyRegistered
+     */
+    public function testPositionDependencyConflict(int $pos, bool $jqAlreadyRegistered): void
+    {
+        $jqAsset = JqueryAsset::class;
+
+        $this->assetManager->setBundles([
+            PositionAsset::class => [
+                'jsOptions' =>  [
+                    'position' => $pos - 1,
+                ],
+            ],
+            JqueryAsset::class => [
+                'jsOptions' =>  [
+                    'position' => $pos,
+                ],
+            ]
+        ]);
+
+        if ($jqAlreadyRegistered) {
+            $message = "An asset bundle that depends on '$jqAsset' has a higher javascript file " .
+            "position configured than '$jqAsset'.";
+
+            $this->expectException(\RuntimeException::class);
+            $this->expectExceptionMessage($message);
+
+            $this->assetManager->register([JqueryAsset::class, PositionAsset::class]);
+        } else {
+            $message = "An asset bundle that depends on '$jqAsset' has a higher javascript file " .
+                "position configured than '$jqAsset'.";
+
+            $this->expectException(\RuntimeException::class);
+            $this->expectExceptionMessage($message);
+
+            $this->assetManager->register([PositionAsset::class]);
+        }
     }
 }
